+++ examples/example-02-ruby-sample/src/layers/ruby.rs
@@ -9,7 +9,7 @@ use libcnb::build::BuildContext;
 use libcnb::data::layer_content_metadata::LayerTypes;
 use libcnb::generic::GenericMetadata;
 use libcnb::layer::{Layer, LayerResult, LayerResultBuilder};
-use libcnb::layer_env::{LayerEnv, ModificationBehavior, TargetLifecycle};
+use libcnb::layer_env::{LayerEnv, ModificationBehavior, Scope};
 
 pub struct RubyLayer;
 
@@ -47,13 +47,13 @@ impl Layer for RubyLayer {
             .env(
                 LayerEnv::new()
                     .chainable_insert(
-                        TargetLifecycle::All,
+                        Scope::All,
                         ModificationBehavior::Prepend,
                         "PATH",
                         context.app_dir.join(".gem/ruby/2.6.6/bin"),
                     )
                     .chainable_insert(
-                        TargetLifecycle::All,
+                        Scope::All,
                         ModificationBehavior::Prepend,
                         "LD_LIBRARY_PATH",
                         layer_path,